+++ Foreign/Nix/Shellout/Prefetch.hs
@@ -19,8 +19,7 @@ module Foreign.Nix.Shellout.Prefetch
   -- * Types
 , PrefetchError(..)
 , Url(..), Sha256(..)
-  -- * Reexports
-, runNixAction, NixAction(..), NixActionError(..)
+, module Foreign.Nix.Shellout.Types
 ) where
 
 import Control.Error hiding (bool, err)
@@ -40,7 +39,7 @@ import Data.Bifunctor (first)
 import qualified Data.Text.Lazy as Text.Lazy
 import qualified Data.Text.Lazy.Encoding as Text.Lazy.Encoding
 import qualified Data.List as List
-import Control.Monad.IO.Class (MonadIO (liftIO))
+import Control.Monad.IO.Class (MonadIO)
 
 data PrefetchError
   = PrefetchOutputMalformed Text
@@ -76,7 +75,7 @@ defaultUrlOptions u = UrlOptions
   , urlExpectedHash = Nothing }
 
 -- | Runs @nix-prefetch-url@.
-url :: (MonadIO m) => UrlOptions -> NixAction m PrefetchError (Sha256, StorePath Realized)
+url :: (MonadIO m) => UrlOptions -> NixAction PrefetchError m (Sha256, StorePath Realized)
 url UrlOptions{..} = Helpers.readProcess handler exec args
   where
     exec = "nix-prefetch-url"
@@ -137,7 +136,7 @@ data GitOutput = GitOutput
   } deriving (Show, Eq)
 
 -- | Runs @nix-prefetch-git@.
-git :: (MonadIO m) => GitOptions -> NixAction m PrefetchError GitOutput
+git :: (MonadIO m) => GitOptions -> NixAction PrefetchError m GitOutput
 git GitOptions{..} = Helpers.readProcess handler exec args
   where
     exec = "nix-prefetch-git"